+++ R/sjPlotGroupFrequencies.R
@@ -98,6 +98,10 @@
 #'          Use \code{"l"} for upper left corner.
 #' @param axis.titles character vector of length one or two, defining the title(s)
 #'          for the x-axis and y-axis.
+#' @param drop.empty Logical, if \code{TRUE} and the variable's values are labelled,
+#'          values that have no observations are still printed in the table (with
+#'          frequency \code{0}). If \code{FALSE}, values / factor levels with no occurence
+#'          in the data are omitted from the output.
 #' @param auto.group numeric value, indicating the minimum amount of unique values
 #'          in the count variable, at which automatic grouping into smaller units
 #'          is done (see \code{\link[sjmisc]{group_var}}). Default value for
@@ -152,6 +156,14 @@
 #' # same data as line plot
 #' sjp.grpfrq(efc$neg_c_7, efc$e42dep, type = "line")
 #'
+#' # show ony categories where we have data (i.e. drop zero-counts)
+#' library(dplyr)
+#' efc <- dplyr::filter(efc, e42dep %in% c(3,4))
+#' sjp.grpfrq(efc$c161sex, efc$e42dep, drop.empty = TRUE)
+#'
+#' # show all categories, even if not in data
+#' sjp.grpfrq(efc$c161sex, efc$e42dep, drop.empty = FALSE)
+#'
 #' @import ggplot2
 #' @importFrom sjstats weight2
 #' @importFrom tidyr gather
@@ -188,6 +200,7 @@ sjp.grpfrq <- function(var.cnt,
                        show.legend = TRUE,
                        show.na = FALSE,
                        show.summary = FALSE,
+                       drop.empty = TRUE,
                        auto.group = NULL,
                        ylim = NULL,
                        grid.breaks = NULL,
@@ -206,6 +219,12 @@ sjp.grpfrq <- function(var.cnt,
   # get variable names
   var.name.cnt <- get_var_name(deparse(substitute(var.cnt)))
   var.name.grp <- get_var_name(deparse(substitute(var.grp)))
+
+  # remove empty value-labels
+  if (drop.empty) {
+    var.cnt <- sjlabelled::drop_labels(var.cnt)
+    var.grp <- sjlabelled::drop_labels(var.grp)
+  }
 
   # copy titles
   if (is.null(axis.titles)) {